// src/common/models/label.group.names.ts
export const POST_VIEW_ACTIONS_LABELS = 'POST_VIEW_ACTIONS_LABELS';
export const POST_VIEW_COMMON_ICONS = 'POST_VIEW_COMMON_ICONS';
export const POST_VIEW_COMMON_LABELS = 'POST_VIEW_COMMON_LABELS';
export const POST_VIEW_POPUP_LABELS = 'POST_VIEW_POPUP_LABELS';

export const DIMMER_LABELS = 'DIMMER_LABELS';

export const enum LABEL_GROUP_NAMES {
    POST_VIEW_ACTIONS_LABELS = 'POST_VIEW_ACTIONS_LABELS',
    POST_VIEW_COMMON_ICONS = 'POST_VIEW_COMMON_ICONS',
    POST_VIEW_COMMON_LABELS = 'POST_VIEW_COMMON_LABELS',
    POST_VIEW_POPUP_LABELS = 'POST_VIEW_POPUP_LABELS',
    DIMMER_LABELS = 'DIMMER_LABELS',
    ERROR_LABELS = 'ERROR_LABELS',
    MENU_LABELS = 'MENU_LABELS',
    MESSAGE_POPUP_LABELS = 'MESSAGE_POPUP_LABELS',
    MESSAGE_COMMON_LABELS = 'MESSAGE_COMMON_LABELS',
    MESSAGE_COMMON_ICONS = 'MESSAGE_COMMON_ICONS',
    MESSAGE_ACTION_LABELS = 'MESSAGE_ACTION_LABELS',
    COMMON_LABELS = 'COMMON_LABELS',
    MESSAGE_DATE_LABELS = 'MESSAGE_DATE_LABELS',
    TRAINING_COMMON_LABELS = 'TRAINING_COMMON_LABELS',
    TRAINING_ERROR_LABELS = 'TRAINING_ERROR_LABLES',
    TRAINING_ACTION_LABLES = 'TRAINING_ACTION_LABLES',
    TRAINING_MESSAGE_LABELS = 'TRAINING_MESSAGE_LABELS',
    PADAWAN_COMMON_LABELS = 'PADAWAN_COMMON_LABELS',
    LOGIN_COMMON_LABELS = 'LOGIN_COMMON_LABELS',
    REGISTRATION_COMMON_LABELS = 'REGISTRATION_COMMON_LABEL',
    USER_COMMON_LABELS = 'USER_COMMON_LABELS',
    USER_MESSAGE_LABELS = 'USER_MESSAGE_LABELS'
}

// Заменить на тип const enum